+++ src/models/barrelConfig.ts
@@ -1,0 +1,10 @@
+export interface EnableBarrelTemplates {
+  header: boolean;
+  footer: boolean;
+}
+
+export interface EnableBarrelMenus {
+  createBarrel: boolean;
+  createDirectoryBarrel: boolean;
+  createFileBarrel: boolean;
+}